// psk-ui/src/app/services/models/question-response.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { Choice } from '../models/choice';
export interface QuestionResponse {
  choices?: Array<Choice>;
  createdBy?: number;
  createdDate?: string;
  id?: number;
  lastModifiedBy?: number;
  lastModifiedDate?: string;
  orderIndex?: number;
  testTemplateId?: number;
  testTemplateTitle?: string;
  text?: string;
  userAnswer?: 'ANSWER_A' | 'ANSWER_B' | 'ANSWER_C' | 'ANSWER_D' | 'ANSWER_E';
}