// psk-ui/src/app/services/models/submit-answer-request.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface SubmitAnswerRequest {
  answerType?: 'ANSWER_A' | 'ANSWER_B' | 'ANSWER_C' | 'ANSWER_D' | 'ANSWER_E';
  choiceId?: number;
  questionId: number;
  textAnswer?: string;
  userTestId: number;
}